impl card check process
