--- src/adapter/mod.rs
@@ -1,1 +1,0 @@
-pub mod mqtt_listener;
