+++ src/domain/repository/mqtt/client.rs
@@ -1,12 +1,13 @@
-use paho_mqtt::{self as mqtt, AsyncReceiver, Message};
+use paho_mqtt::{self as mqtt, AsyncClient, AsyncReceiver, Message};
 use std::collections::HashMap;
 use std::sync::Arc;
+use crate::server::connection::RequestContext;
 
 pub trait MqttClientRepository {
     async fn connect(&mut self) -> anyhow::Result<()>;
     fn disconnect(&self) -> anyhow::Result<()>;
     fn subscribe(&mut self, topic: &str, handler: MessageHandler) -> Result<(), mqtt::Error>;
-    fn publish(&mut self, topic: &str, message: &str) -> Result<(), mqtt::Error>;
+    fn publish(&self, topic: &str, message: &str) -> anyhow::Result<()>;
 
     fn get_connection(&self) -> &paho_mqtt::AsyncClient;
 
@@ -17,6 +18,6 @@ pub trait MqttClientRepository {
     // fn get_stream(&mut self) -> &AsyncReceiver<Option<Message>>;
 }
 
-pub type MessageHandler = Arc<dyn Fn(&Message) + Send + Sync>;
+pub type MessageHandler = Arc<dyn Fn(&AsyncClient, &Message, &RequestContext) + Send + Sync>;
 
-pub trait MqttListener: Fn(Message) + Send + Sync + 'static {}
+pub trait MqttListener: Fn(AsyncClient, Message) + Send + Sync + 'static {}
