+++ src/server/mod.rs
@@ -1,5 +1,5 @@
-mod connection;
-mod handler;
+pub mod connection;
+pub mod handler;
 pub mod request;
 mod response;
 pub mod router;
